+++ src/main/java/dev/smootheez/elytracontrol/mixin/OptionsMixin.java
@@ -1,26 +1,28 @@
 package dev.smootheez.elytracontrol.mixin;
 
-import dev.smootheez.elytracontrol.*;
+import dev.smootheez.elytracontrol.registry.*;
 import net.minecraft.client.*;
 import org.apache.commons.lang3.*;
 import org.spongepowered.asm.mixin.*;
 import org.spongepowered.asm.mixin.injection.*;
 import org.spongepowered.asm.mixin.injection.callback.*;
 
-import java.io.*;
-
 @Mixin(Options.class)
-public class OptionsMixin {
+public abstract class OptionsMixin {
     @Mutable
     @Final
     @Shadow
     public KeyMapping[] keyMappings;
 
-    @Inject(method = "<init>", at = @At("TAIL"))
-    private void addCustomKeybind(Minecraft minecraft, File file, CallbackInfo info) {
-        this.keyMappings = ArrayUtils.add(
+    @Inject(method = "<init>", at = @At(
+            value = "INVOKE",
+            target = "Lnet/minecraft/client/Options;load()V"
+    ))
+    private void addCustomKeybind(CallbackInfo info) {
+        this.keyMappings = ArrayUtils.addAll(
                 this.keyMappings,
-                KeyBindsRegistry.CONTROL_FLYING
+                KeyBindsRegistry.START_FLYING,
+                KeyBindsRegistry.STOP_FLYING
         );
     }
 }