+++ src/main/java/dev/smootheez/elytracontrol/mixin/PlayerEntityMixin.java
@@ -4,5 +4,5 @@ import net.minecraft.world.entity.player.*;
 import org.spongepowered.asm.mixin.*;
 
 @Mixin(Player.class)
-public class PlayerEntityMixin {
+public abstract class PlayerEntityMixin {
 }